refactor: Move public to src folder
