+++ src/public/assets/works/dini-notebook/IMG_2.png
