+++ src/public/assets/works/dini-notebook/IMG_6.png
